Refactor profile to use IepDialog component
